requirements.txt
torch>=1.9.0,<2.0.0
torchvision>=0.10.0,<0.15.0
transformers>=4.20.0,<4.30.0
gradio>=3.50.0,<4.0.0
Pillow>=8.0.0,<10.0.0
opencv-python>=4.5.0,<4.8.0
albumentations>=1.3.0,<1.4.0
numpy>=1.21.0,<1.25.0
scikit-learn>=1.0.0,<1.3.0
datasets>=2.0.0,<2.14.0
accelerate>=0.20.0,<0.21.0
huggingface-hub>=0.15.0,<0.16.0 